@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\Views\AuditoriumListWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "BDDC9C1D701BBFF02E04F8B077EAEC4274FB4242"
+﻿#pragma checksum "..\..\..\..\Views\AuditoriumListWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "4B4E156A5A3F690C3417094C51528487E1349506"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
