@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\Views\MovieInformationWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "3FB047DA2D31E37A72C1F01FCF8F47EBB78BE617"
+﻿#pragma checksum "..\..\..\..\Views\MovieInformationWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "27737883E9C9AD601DE09D721E36607A14DE5569"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
@@ -32,7 +32,7 @@ using System.Windows.Shapes;
 using System.Windows.Shell;
 
 
-namespace TheaterManagementApp.Views {
+namespace Theater_Management_FE.Views {
     
     
     /// <summary>
