@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\Views\MovieListWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "29CE9D9C1B15BD792BF81197E608EA4874425B73"
+﻿#pragma checksum "..\..\..\..\Views\MovieListWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "8C57BA5B0356D23F12499D516CD7C7E85BE5C341"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
@@ -32,7 +32,7 @@ using System.Windows.Shapes;
 using System.Windows.Shell;
 
 
-namespace TheaterManagementApp.Views {
+namespace Theater_Management_FE.Views {
     
     
     /// <summary>
