@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\Views\ProfileWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "EA1A33141256CBAE54BD44658C4944EB116F3FEC"
+﻿#pragma checksum "..\..\..\..\Views\ProfileWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "A63C1AE4826083339BC616877098436A85916EC3"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
@@ -32,7 +32,7 @@ using System.Windows.Shapes;
 using System.Windows.Shell;
 
 
-namespace TheaterManagementApp.Views {
+namespace Theater_Management_FE.Views {
     
     
     /// <summary>
