@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\Views\ShowtimeListWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "05B1068CDE21CC9E1182753B4F8E25D484F3FC78"
+﻿#pragma checksum "..\..\..\..\Views\ShowtimeListWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "82809CE528FD131ECCD1FCDCBC363CB3CEAE41BD"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
@@ -32,7 +32,7 @@ using System.Windows.Shapes;
 using System.Windows.Shell;
 
 
-namespace TheaterManagementApp.Views {
+namespace Theater_Management_FE.Views {
     
     
     /// <summary>
