@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\Views\SignInWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "7B28ACB30C7A0BC64E4C7A5D032491C25DA9D902"
+﻿#pragma checksum "..\..\..\..\Views\SignInWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "283D784E00654A9D443455E24D547F453A1A684A"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
@@ -32,7 +32,7 @@ using System.Windows.Shapes;
 using System.Windows.Shell;
 
 
-namespace TheaterManagementApp.Views {
+namespace Theater_Management_FE.Views {
     
     
     /// <summary>
