@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\Views\SignUpWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "E2558D66DC9997D2EDB4BEEDE014EE7C42017D78"
+﻿#pragma checksum "..\..\..\..\Views\SignUpWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "9626BB60FEA8CDB109102EDED2C0A2621EF3102F"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
@@ -32,7 +32,7 @@ using System.Windows.Shapes;
 using System.Windows.Shell;
 
 
-namespace TheaterManagementApp.Views {
+namespace Theater_Management_FE.Views {
     
     
     /// <summary>
