@@ -1,4 +1,4 @@
-﻿#pragma checksum "..\..\..\..\Views\TinTucWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "0A210E1C8342FEA315A4FCB22BC39DE71BA87B3F"
+﻿#pragma checksum "..\..\..\..\Views\TinTucWindow.xaml" "{ff1816ec-aa5e-4d10-87f7-6f4963833460}" "BCD084DC37B97C1F7BDDAB5B9F9848548AF0B126"
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
@@ -32,7 +32,7 @@ using System.Windows.Shapes;
 using System.Windows.Shell;
 
 
-namespace TheaterManagementApp.Views {
+namespace Theater_Management_FE.Views {
     
     
     /// <summary>
